+++ mex/common/models/resource.py
@@ -7,7 +7,12 @@ from pydantic import Field, computed_field
 from mex.common.models.base import BaseModel
 from mex.common.models.extracted_data import ExtractedData
 from mex.common.models.merged_item import MergedItem
-from mex.common.models.rules import AdditiveRule, PreventiveRule, SubtractiveRule
+from mex.common.models.rules import (
+    AdditiveRule,
+    PreventiveRule,
+    RuleSet,
+    SubtractiveRule,
+)
 from mex.common.types import (
     AccessRestriction,
     AnonymizationPseudonymization,
@@ -370,3 +375,14 @@ class PreventiveResource(_Stem, PreventiveRule):
     title: list[MergedPrimarySourceIdentifier] = []
     unitInCharge: list[MergedPrimarySourceIdentifier] = []
     wasGeneratedBy: list[MergedPrimarySourceIdentifier] = []
+
+
+class ResourceRuleSet(_Stem, RuleSet):
+    """Set of rules to edit a resource item."""
+
+    entityType: Annotated[
+        Literal["ResourceRuleSet"], Field(alias="$type", frozen=True)
+    ] = "ResourceRuleSet"
+    additive: AdditiveResource
+    subtractive: SubtractiveResource
+    preventive: PreventiveResource