+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `rki/mex` user agent is sent with query requests via wikidata connector
 
 ### Changes
+- changed backend api connector payload to "items"
 
 - update wikidata search organization request query, with optional language parameter
   wikidata query search can be enhanced by specifying the language.